refactor: speed up cold Google policy checks
Share the existing Gemini classifiers through a lightweight AI entry so policy checks avoid streaming imports. Preserve historical SDK names and thinking/replay behavior; retain the existing policy cases with an import-boundary regression guard.
