+++ extensions/google/thinking-api.ts
@@ -1,11 +1,13 @@
+export {
+  isGoogleGemini3FlashModel,
+  isGoogleGemini3ProModel,
+  isGoogleGemini3ThinkingLevelModel,
+} from "openclaw/plugin-sdk/provider-thinking-runtime";
 // Google API module exposes the plugin public contract.
 export {
   createGoogleThinkingPayloadWrapper,
   createGoogleThinkingStreamWrapper,
   isGoogleGemini25ThinkingBudgetModel,
-  isGoogleGemini3FlashModel,
-  isGoogleGemini3ProModel,
-  isGoogleGemini3ThinkingLevelModel,
   isGoogleThinkingRequiredModel,
   resolveGoogleGemini3ThinkingLevel,
   sanitizeGoogleThinkingPayload,
