+++ packages/ai/README.md
@@ -23,7 +23,7 @@ Host policy (request fetch guarding, secret redaction, strict-tool defaults,
 provider plugin hooks, and diagnostics logging) can be injected with
 `configureAiTransportHost`; the defaults are inert.
 
-The explicit `@openclaw/ai/internal/anthropic`, `openai`,
+The explicit `@openclaw/ai/internal/anthropic`, `google-model-family`, `openai`,
 `openai-responses-payload-policy`, `retry-after`, `runtime`, `shared`, and
 `tool-schema` subpaths exist for the OpenClaw application itself.
 They carry no semver guarantee and can change or disappear in any release; do
